move repo dropdown to header, reorganize components
